cfg: Remove workaround for "httptpl" with missing ending slash.
There are a lot of paths described in the config and it makes
little sense to fixup (add / at the end) only for "httptpl" and not
for the others.

The users must set the directory with slash (/) at the end in
order for that directory to be correctly prepended to file names.

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@7396 4b0bc96b-bc66-0410-9d44-ebda105a78c1