Mark all cs_malloc() calls failable.
Instead of killing OSCam if an allocation fails, it is better to
try to handle it. Most of the cs_malloc() calls were handling
allocations gracefully but there were a lot that just killed the
server.

Now all allocations are checked. If you run out of memory that will
not help you much, but at least OSCam will not crash without any
explanation.

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@7596 4b0bc96b-bc66-0410-9d44-ebda105a78c1