Try to fix OK and ERROR definitions mess.
We have shot ourselves in the foot by using OK and ERROR defines.
In csctapi have

  OK    0
  ERROR 1

but card systems are using the exact opposite

  OK    1
  ERROR 0

On top of that CYGWIN have its own set of OK and ERROR defines.

This tries to fix ticket #2850 and cure a regression accidently
introduced in r7688 by moving some defines around.

After this OK and ERROR in csctapi are defined in each C file that
uses them so it is hard for them to escape outside of the file,
also OK and ERROR are undefined in case they are already defined.

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@7697 4b0bc96b-bc66-0410-9d44-ebda105a78c1