webif: Show picon image only when the picon exists.
It turns out that showing empty image in the place of non existing
picon annoys users (who would've known :) With this changeset the
code checks whether picon can be found and only if it exists then
shows it, otherwise channel name is shown instead.

This is another fix for ticket #2672.

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@7735 4b0bc96b-bc66-0410-9d44-ebda105a78c1