writelock's reader based not module based anymore
Now the write lock's for reader init and settings are module based, not static anymore and so blocking the read to other smartreaders as well during init fase.
The cd card detect does not work for tripple (it did not on slow devices anyway with the new way). Working on a solution trough read call back for triple ass well.

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@9000 4b0bc96b-bc66-0410-9d44-ebda105a78c1