Changed baud rate settings
The baud rate setting only needed at reader setup and card init. After this the used baudrate is determined by mhz used. So just set the ftdi to max avbl speeds which is 3000000 for this device. Since the line break issue is done setting the correct actual used baud is not important anymore.

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@9021 4b0bc96b-bc66-0410-9d44-ebda105a78c1