dvbapi: add adapter index for BOXTYPE_PC socket messages
When CA_SET_PID was received on single socket a client
was unable to distinguish between adapters, and this
lead to problems when decoding eg. same PID numbers on
different adapters.
With this additional information a client can use
a single socket connection for multiple adapters.

This small protocol change was consulted with client
developers.


git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@9095 4b0bc96b-bc66-0410-9d44-ebda105a78c1