dvbapi: Add support for filtering over network
This commit adds possibility to use dvbapi over a TCP network connection
instead of local camd.socket.
Starting filters and obtaining ECM/CAT/EMM data is done on client,
this way no real demux hardware is needed. This way the boxes like
SAT>IP can be handled when the dvbapi client supports this mode.
Additionally this change would allow to get rid one local OSCam
(which was needed because the real hardware was necessary) and connect
directly to OSCam with card reader.

Additional 'listen_port' option was added to enable this mode.


git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@9574 4b0bc96b-bc66-0410-9d44-ebda105a78c1