Added new parameter in global section (oscam.conf):
nds_swapp_cw = 0|1 (deafult=0)
When 0, it means that NDS cw's will be checked, and if they don't respect odd/even parts, they are dropped. No pushed with cacheex and not served with "normal" readers.
More, using this =0, we force cache to check only cw's with right odd/even parts, prevent it to get the swapped cw (if present).

When 1, behaviour like previous oscam's.


git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@9994 4b0bc96b-bc66-0410-9d44-ebda105a78c1