Do not work directly with stb_boxtype.
stb_boxtype is possible to be NULL, in which case using strcasecmp()
or similar functions would cause segfault. This patch defines two new
functions for dealing with stb_boxtype.

  boxtype(X) checks if stb_boxtype equals X.
  boxtype_get() returns the detected box type string.

As a nice side effect (in addition to cleaner and safer code) we shave
off a bit of binary size.

bloatcheck reports:
{{{
 add/remove: 2/0 grow/shrink: 0/3 up/down: 63/-255 (-192)
 function                                     old     new   delta
 boxtype                                        -      43     +43
 boxtype_get                                    -      20     +20
 dvbapi_ioctl                                 367     359      -8
 Sci_Deactivate                               251     235     -16
 cardreader_init                             1784    1553    -231
}}}

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10467 4b0bc96b-bc66-0410-9d44-ebda105a78c1