Newcamd:
- Fix for non-matching emm caid (introduced r10582)

However the real reason is that before it was broken too: emmcaid was always 0000 but it was eventually checked against 0000 so it passed
The improvements gf introduced made the emmchecks more strict and so the bad code of newcamd started to fail.

Tnx Dukynukem for not whining and complaining like others did but eventually providing 2 simple logs and testing the patch!


git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10630 4b0bc96b-bc66-0410-9d44-ebda105a78c1